app.lifecycle subcapability treeshaking refactor
